added about saudi OSS
