+++ homepage/public/index.html
@@ -138,6 +138,29 @@
           المجموعة السعودية للمصادر المفتوحة
         </p>
       </div>
+      <div class="connector">
+        <blockquote>
+          <p>
+            الهدف من فكرة جمع المشاريع المفتوحة المصدر هو تحفيز المجتمع التقني
+            السعودي لإثراء المحتوى البرمجي وخلق بيئة تعارف بين المبرمجين من خلال
+            عرض بعض من أعمالهم. هذه المجموعة تدار من قبل
+            <a
+              target="_blank"
+              href="https://github.com/orgs/SaudiOpenSourceCommunity/people"
+            >
+              متطوعين.
+            </a>
+            إذا أردت أن تضيف مشروعك هنا, الرجاء زيارة حساب
+            <a
+              target="_blank"
+              href="https://github.com/SaudiOpenSourceCommunity/SaudiOSS"
+            >
+              المجموعة.
+            </a>
+          </p>
+        </blockquote>
+        <div class="devs-title">المبرمجون</div>
+      </div>
       <div id="root">
         <div class="loader"></div>
         <div>
