+++ homepage/src/App.js
@@ -1,66 +1,16 @@
-import React, { useState, useEffect, useRef } from "react";
+import React, { useState, useEffect } from "react";
 import {
   Card,
   Button,
   Avatar,
   Modal,
   Badge,
-  ButtonIcon,
-  Spinner,
   Pagination
 } from "react-rainbow-components";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCode } from "@fortawesome/free-solid-svg-icons";
-import shuffle from "shuffle-array";
 import "./App.css";
-
-const githubURL =
-  "https://raw.githubusercontent.com/SaudiOpenSourceCommunity/SaudiOSS/master/README.md";
-
-function parseHTML(html) {
-  var t = document.createElement("template");
-  t.innerHTML = html;
-  return t.content.cloneNode(true);
-}
-
-async function fetchProjects() {
-  let openSourceProjects = [];
-  openSourceProjects = await fetch(githubURL).then(payload => payload.text());
-  return tableToJSON(openSourceProjects);
-}
-
-//This looks ugly 🙁 but could not figure out better way to do it.
-function tableToJSON(table) {
-  const developers = [];
-  const TableNode = Array.from(
-    parseHTML(table).querySelector("tbody").children
-  ).splice(1);
-  for (let i = 0; i < TableNode.length; i++) {
-    let row = TableNode[i].cells;
-    const developer = {};
-    developer.NumberOfProjects = row[0].attributes[0].value;
-    developer.name = row[0].innerText.trim();
-    developer.githubURL = row[0].children[0]
-      ? row[0].children[0].href.trim()
-      : undefined;
-    developer.projects = [];
-    developer.projects.push({
-      name: row[1].innerText.trim(),
-      URL: row[1].children[0].href.trim(),
-      description: row[2].innerText.trim()
-    });
-    for (let j = 1; j < developer.NumberOfProjects; j++) {
-      row = TableNode[++i].cells;
-      developer.projects.push({
-        name: row[0].innerText.trim(),
-        URL: row[0].children[0].href.trim(),
-        description: row[1].innerText.trim()
-      });
-    }
-    developers.push(developer);
-  }
-  return shuffle(developers);
-}
+import fetchProjects from "./githubApi";
 
 function App() {
   const [developers, loadDevelopers] = useState([]);
@@ -71,7 +21,6 @@ function App() {
     };
     fetchDev();
   }, []);
-  console.log(developers);
   if (developers.length === 0) return <></>;
   return (
     <div className="devs-container">
@@ -87,15 +36,12 @@ function DevInfo({ developer }) {
   const [page, pageChange] = useState(1);
 
   const getProjectContent = () => {
-    console.log(developer, page);
     const project = developer.projects[page - 1];
     return (
       <>
         <div>
-          {" "}
-          <a href={project.URL}>
-            {" "}
-            <b style={{ fontSize: "1.5em" }}>{project.name}</b>{" "}
+          <a target="_blank" href={project.URL}>
+            <b style={{ fontSize: "1.5em" }}>{project.name}</b>
           </a>
         </div>
         <div>{project.description}</div>
@@ -107,7 +53,9 @@ function DevInfo({ developer }) {
     if (developer.githubURL)
       return (
         <h1 className="rainbow-p-top_large rainbow-font-size-heading_small rainbow-color_dark-1">
-          <a href={developer.githubURL}>{developer.name}</a>
+          <a target="_blank" href={developer.githubURL}>
+            {developer.name}
+          </a>
         </h1>
       );
     return (
